Implement the string size operation with size_t
The smart string only declared a size function,
but never defined it in the implementation. So,
the function headers were changed to return
size_t (the right type) instead of unsigned int.
(As is done in the C++ standard for strings).

sstring.h:
 * fix the string_size and string_length
   function header return types

sstring.c:
 * implement string_size.

+++ sstring.h
@@ -48,12 +48,12 @@ string_empty
 (struct string *s);
 
 // size
-unsigned int
+size_t
 string_size
 (struct string *s);
 
 // length
-unsigned int
+size_t
 string_length
 (struct string *s);
